backend_furia/requirements.txt
python-telegram-bot==20.7
wikipedia-api
requests
